# docs/SmartHome_Swagger.yaml
openapi: 3.0.3
info:
  title: SmartHome
  description: Документация к системе "Умный дом"
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Локальный сервер
...
paths:
...
  /devices:
    post:
      tags: 
        - device
      summary: Создание нового устройства (подключение)
      operationId: createDevice
      parameters:
        - in: path
          name: deviceId
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Device'
            examples:
              example:
                summary: Пример
                value:
                  house_id: 1
                  type_id: 1
                  serial_number: "12345qwerty"
                  status: "off"
                  is_active: true
                  created_at: "2025-01-01T00:00:00Z"
      responses:
        '200':
          description: Устройство создано
        '404':
          description: Устройство не найдено
...
components:
  schemas:
...
    Device:
      type: object
      properties:
        id:
          type: number
        house_id:
          type: number
        type_id:
          type: number
        serial_number:
          type: string
        status:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time